requirements.txt
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
fastapi==0.92.0
Flask==2.2.2
greenlet==2.0.2
itsdangerous==2.1.2
orm==0.3.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.46
starlette==0.25.0
typesystem==0.3.1
Werkzeug==2.2.2
passlib                # for password hashing
pydantic[email]
python-multipart